rename getRepositoryUrl to get-repository-url
